refactor: move delete button to general settings
